+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Improve performance of ASCII rendering - [@dblock](https://github.com/dblock).
 * Render ASCII images on a background thread - [@dblock](https://github.com/dblock).
 * Use "Courier New" font instead of "Courier" - [@dblock](https://github.com/dblock).
+* Don't scale image when `UIImageView.contentMode` is set to `UIViewContentModeScaleToFill` - [@dblock](https://github.com/dblock).
 
 ## 1.0.0 (3/21/2014)